+++ package.json
@@ -15,13 +15,13 @@
     "html-loader": "^5.1.0",
     "html-webpack-plugin": "^5.6.4",
     "style-loader": "^4.0.0",
-    "webpack": "^5.101.3",
+    "webpack": "^5.102.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.15.2",
     "webpack-glsl-loader": "^1.0.1"
   },
   "dependencies": {
-    "@kitware/vtk.js": "^32.13.0",
+    "@kitware/vtk.js": "^32.14.0",
     "@tensorflow/tfjs": "^4.22.0",
     "three": "^0.175.0",
     "vtk.js": "^24.19.2",
